bindings: images uses entities/types
