+++ pkg/domain/entities/images.go
@@ -3,15 +3,13 @@ package entities
 import (
 	"io"
 	"net/url"
-	"time"
 
 	"github.com/containers/common/pkg/config"
 	"github.com/containers/image/v5/manifest"
 	"github.com/containers/image/v5/signature/signer"
 	"github.com/containers/image/v5/types"
 	encconfig "github.com/containers/ocicrypt/config"
-	"github.com/containers/podman/v4/pkg/inspect"
-	"github.com/containers/podman/v4/pkg/trust"
+	entitiesTypes "github.com/containers/podman/v4/pkg/domain/entities/types"
 	"github.com/docker/docker/api/types/container"
 	"github.com/opencontainers/go-digest"
 	v1 "github.com/opencontainers/image-spec/specs-go/v1"
@@ -54,37 +52,7 @@ func (i *Image) Id() string { //nolint:revive,stylecheck
 }
 
 // swagger:model LibpodImageSummary
-type ImageSummary struct {
-	ID          string `json:"Id"`
-	ParentId    string //nolint:revive,stylecheck
-	RepoTags    []string
-	RepoDigests []string
-	Created     int64
-	Size        int64
-	SharedSize  int
-	VirtualSize int64
-	Labels      map[string]string
-	Containers  int
-	ReadOnly    bool `json:",omitempty"`
-	Dangling    bool `json:",omitempty"`
-
-	// Podman extensions
-	Names   []string `json:",omitempty"`
-	Digest  string   `json:",omitempty"`
-	History []string `json:",omitempty"`
-}
-
-func (i *ImageSummary) Id() string { //nolint:revive,stylecheck
-	return i.ID
-}
-
-func (i *ImageSummary) IsReadOnly() bool {
-	return i.ReadOnly
-}
-
-func (i *ImageSummary) IsDangling() bool {
-	return i.Dangling
-}
+type ImageSummary = entitiesTypes.ImageSummary
 
 // ImageRemoveOptions can be used to alter image removal.
 type ImageRemoveOptions struct {
@@ -102,30 +70,12 @@ type ImageRemoveOptions struct {
 
 // ImageRemoveReport is the response for removing one or more image(s) from storage
 // and images what was untagged vs actually removed.
-type ImageRemoveReport struct {
-	// Deleted images.
-	Deleted []string `json:",omitempty"`
-	// Untagged images. Can be longer than Deleted.
-	Untagged []string `json:",omitempty"`
-	// ExitCode describes the exit codes as described in the `podman rmi`
-	// man page.
-	ExitCode int
-}
+type ImageRemoveReport = entitiesTypes.ImageRemoveReport
 
 type ImageHistoryOptions struct{}
 
-type ImageHistoryLayer struct {
-	ID        string    `json:"id"`
-	Created   time.Time `json:"created,omitempty"`
-	CreatedBy string    `json:",omitempty"`
-	Tags      []string  `json:"tags,omitempty"`
-	Size      int64     `json:"size"`
-	Comment   string    `json:"comment,omitempty"`
-}
-
-type ImageHistoryReport struct {
-	Layers []ImageHistoryLayer
-}
+type ImageHistoryLayer = entitiesTypes.ImageHistoryLayer
+type ImageHistoryReport = entitiesTypes.ImageHistoryReport
 
 // ImagePullOptions are the arguments for pulling images.
 type ImagePullOptions struct {
@@ -166,16 +116,7 @@ type ImagePullOptions struct {
 }
 
 // ImagePullReport is the response from pulling one or more images.
-type ImagePullReport struct {
-	// Stream used to provide output from c/image
-	Stream string `json:"stream,omitempty"`
-	// Error contains text of errors from c/image
-	Error string `json:"error,omitempty"`
-	// Images contains the ID's of the images pulled
-	Images []string `json:"images,omitempty"`
-	// ID contains image id (retained for backwards compatibility)
-	ID string `json:"id,omitempty"`
-}
+type ImagePullReport = entitiesTypes.ImagePullReport
 
 // ImagePushOptions are the arguments for pushing images.
 type ImagePushOptions struct {
@@ -261,14 +202,7 @@ type ImagePushReport struct {
 
 // ImagePushStream is the response from pushing an image. Only used in the
 // remote API.
-type ImagePushStream struct {
-	// ManifestDigest is the digest of the manifest of the pushed image.
-	ManifestDigest string `json:"manifestdigest,omitempty"`
-	// Stream used to provide push progress
-	Stream string `json:"stream,omitempty"`
-	// Error contains text of errors from pushing
-	Error string `json:"error,omitempty"`
-}
+type ImagePushStream = entitiesTypes.ImagePushStream
 
 // ImageSearchOptions are the arguments for searching images.
 type ImageSearchOptions struct {
@@ -296,22 +230,7 @@ type ImageSearchOptions struct {
 }
 
 // ImageSearchReport is the response from searching images.
-type ImageSearchReport struct {
-	// Index is the image index (e.g., "docker.io" or "quay.io")
-	Index string
-	// Name is the canonical name of the image (e.g., "docker.io/library/alpine").
-	Name string
-	// Description of the image.
-	Description string
-	// Stars is the number of stars of the image.
-	Stars int
-	// Official indicates if it's an official image.
-	Official string
-	// Automated indicates if the image was created by an automated build.
-	Automated string
-	// Tag is the repository tag
-	Tag string
-}
+type ImageSearchReport = entitiesTypes.ImageSearchReport
 
 // Image List Options
 type ImageListOptions struct {
@@ -329,9 +248,7 @@ type ImageTagOptions struct{}
 type ImageUntagOptions struct{}
 
 // ImageInspectReport is the data when inspecting an image.
-type ImageInspectReport struct {
-	*inspect.ImageData
-}
+type ImageInspectReport = entitiesTypes.ImageInspectReport
 
 type ImageLoadOptions struct {
 	Input           string
@@ -339,9 +256,7 @@ type ImageLoadOptions struct {
 	SignaturePolicy string
 }
 
-type ImageLoadReport struct {
-	Names []string
-}
+type ImageLoadReport = entitiesTypes.ImageLoadReport
 
 type ImageImportOptions struct {
 	Architecture    string
@@ -356,9 +271,7 @@ type ImageImportOptions struct {
 	SourceIsURL     bool
 }
 
-type ImageImportReport struct {
-	Id string //nolint:revive,stylecheck
-}
+type ImageImportReport = entitiesTypes.ImageImportReport
 
 // ImageSaveOptions provide options for saving images.
 type ImageSaveOptions struct {
@@ -413,9 +326,7 @@ type ImageTreeOptions struct {
 }
 
 // ImageTreeReport provides results from ImageEngine.Tree()
-type ImageTreeReport struct {
-	Tree string // TODO: Refactor move presentation work out of server
-}
+type ImageTreeReport = entitiesTypes.ImageTreeReport
 
 // ShowTrustOptions are the cli options for showing trust
 type ShowTrustOptions struct {
@@ -426,12 +337,7 @@ type ShowTrustOptions struct {
 }
 
 // ShowTrustReport describes the results of show trust
-type ShowTrustReport struct {
-	Raw                     []byte
-	SystemRegistriesDirPath string
-	JSONOutput              []byte
-	Policies                []*trust.Policy
-}
+type ShowTrustReport = entitiesTypes.ShowTrustReport
 
 // SetTrustOptions describes the CLI options for setting trust
 type SetTrustOptions struct {
@@ -466,18 +372,10 @@ type ImageUnmountOptions struct {
 }
 
 // ImageMountReport describes the response from image mount
-type ImageMountReport struct {
-	Id           string //nolint:revive,stylecheck
-	Name         string
-	Repositories []string
-	Path         string
-}
+type ImageMountReport = entitiesTypes.ImageMountReport
 
 // ImageUnmountReport describes the response from umounting an image
-type ImageUnmountReport struct {
-	Err error
-	Id  string //nolint:revive,stylecheck
-}
+type ImageUnmountReport = entitiesTypes.ImageUnmountReport
 
 const (
 	LocalFarmImageBuilderName   = "(local)"
@@ -485,10 +383,4 @@ const (
 )
 
 // FarmInspectReport describes the response from farm inspect
-type FarmInspectReport struct {
-	NativePlatforms   []string
-	EmulatedPlatforms []string
-	OS                string
-	Arch              string
-	Variant           string
-}
+type FarmInspectReport = entitiesTypes.FarmInspectReport
