+++ pkg/domain/entities/types/types.go
@@ -1,6 +1,12 @@
 // copied from github.com/docker/docker/api/types
 package types
 
+import (
+	"os"
+
+	buildahDefine "github.com/containers/buildah/define"
+)
+
 // ComponentVersion describes the version information for a specific component.
 type ComponentVersion struct {
 	Name    string
@@ -41,4 +47,33 @@ type ContainerCreateResponse struct {
 	// Warnings during container creation
 	// required: true
 	Warnings []string `json:"Warnings"`
+}
+
+// FarmBuildOptions describes the options for building container images on farm nodes
+type FarmBuildOptions struct {
+	// Cleanup removes built images from farm nodes on success
+	Cleanup bool
+	// Authfile is the path to the file holding registry credentials
+	Authfile string
+	// SkipTLSVerify skips tls verification when set to true
+	SkipTLSVerify bool
+}
+
+// BuildOptions describe the options for building container images.
+type BuildOptions struct {
+	buildahDefine.BuildOptions
+	ContainerFiles []string
+	FarmBuildOptions
+	// Files that need to be closed after the build
+	// so need to pass this to the main build functions
+	LogFileToClose *os.File
+	TmpDirToClose  string
+}
+
+// BuildReport is the image-build report.
+type BuildReport struct {
+	// ID of the image.
+	ID string
+	// Format to save the image in
+	SaveFormat string
 }
